requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-environ>=0.11
apscheduler>=3.10
passlib[bcrypt]>=1.7
jinja2>=3.1
python-multipart>=0.0.9
gunicorn>=21.0
django>=4.2,<5.0
djangorestframework>=3.14
django-environ>=0.11
apscheduler>=3.10
passlib[bcrypt]>=1.7
jinja2>=3.1
python-multipart>=0.0.9
gunicorn>=21.0
psycopg2-binary>=2.9   
